requirements.txt
alembic==1.14.0
dateparser==1.2.2
fastapi==0.118.0
pandas==2.3.3
pydantic==2.12.0
python-decouple==3.8
SQLAlchemy==2.0.36
starlette==0.48.0
uvicorn==0.37.0
yfinance==0.2.66
psycopg2-binary==2.9.10
python-multipart==0.0.20
mcp==1.16.0
fastmcp==2.12.4
python-slugify==8.0.4
redis==6.4.0